requirements.txt
# requiremnts.txt

sqlalchemy
sqlalchemy_mptt
python-settings
python-dotenv
